requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
streamlit>=1.44,<2.0
requests>=2.32,<3.0
Pillow>=10.4,<12.0
python-multipart>=0.0.9,<1.0
opencv-python>=4.10,<5.0
torch>=2.4,<3.0
torchvision>=0.19,<1.0
diffusers>=0.32,<1.0
transformers>=4.44,<5.0
safetensors>=0.4,<1.0
accelerate>=1.0,<2.0
